Added interim kafka topic where messages from original topic are flattened.  Added proper batch saving to db, introduced kafka offset for each message.

+++ src/main/java/ru/dmzadorin/demo/model/Message.java
@@ -1,14 +1,14 @@
 package ru.dmzadorin.demo.model;
 
 public class Message {
-    private Integer messageId;
+    private Long messageId;
     private String payload;
 
-    public Integer getMessageId() {
+    public Long getMessageId() {
         return messageId;
     }
 
-    public void setMessageId(Integer messageId) {
+    public void setMessageId(Long messageId) {
         this.messageId = messageId;
     }
